requirements.txt
pandas==1.5.3
numpy==1.24.0
matplotlib==3.6.3
networkx==2.8.8
pyyaml==6.0.2
openpyxl==3.0.10
plotly==5.3.0
python-louvain
scipy==1.9.3
pyvis
seaborn==0.11.2